Don't connect dots that are too far apart
Disable interpolation when we detect a time gap in the data.

Add a simple initial heuristic for detecting gaps. Compute all the
time-deltas, find the 75th percentile, and multiply it by a factor
(currently 1.5). The result is the gap threshold. This heuristic will
be easy to tweak now that it's in place.

We have to detect gaps in aggregated and unaggregated data
independently, and then merge them. This was difficult to fit into the
existing code, so I shuffled the way we compute the Dygraphs input data.