Fixes integration test and the way we have of exporting files.
The convention (introduced by the PR #751 by @mrcslws and which fixes UNI-334) is that we allow parsing of date/time values that have a timezone, but that we do not take the timezone into account in the app for vizualization or export. This means that a CSV that has a timezone can be parsed but that the timzone will not be represented in the app (i.e. timezones are not reflected in the charts or in the exported result CSV)